Add concrete parsers.
AircraftTypesCountParser - Parses information about the types of aircraft that flights were made on and the number of flights for each type.
DaysWithMultipleFlightsParser - Parses information about days with more than 1 flight.
TotalFlightTimeParser - Parses information about the total amount of time spent in air.
UniqueDaysWithFlightsParser - Parses information about unique days on which flights took place.
YearParser - Parses the year in which the logbook was created.